+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.4.0
+Features:
+    - Command line parsing is now done internally, i.e. removed optparse (#6).
+    
+Bug fix:
+    - OptParse was hijacking the -v --version.
+    
 ## 0.3.0
 Features:
     - Added a shell class for any user interaction.